Revert "コミット"
This reverts commit c143ee7eee2d86c2827cded0df3c98a1954abe1d.

+++ db/migrate/20231219031459_devise_create_users.rb
@@ -8,9 +8,9 @@ class DeviseCreateUsers < ActiveRecord::Migration[7.0]
       t.string  :email,              null: false, default: ""
       t.string  :encrypted_password, null: false, default: ""
       t.date    :birthday,           null: false
-      t.integer :sex_id,             null: false
-      t.float   :height,             null: false
-      t.float   :weight,             null: false
+      t.integer :sex_id,                null: false
+      t.integer :height,             null: false
+      t.integer :weight,             null: false
 
       ## Recoverable
       t.string   :reset_password_token
